deployment/requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
Pillow
plotly
scikit-learn == 1.4.0